// apps/pc/src/generated/models/SystemusersModel.ts
export const Systemusersaccessmode = {
  0: '読み取り/書き込み',
  1: '管理',
  2: '読み取り',
  3: 'サポート ユーザー',
  4: '非対話型',
  5: '代理管理者'
} as const;
export type Systemusersaccessmode = keyof typeof Systemusersaccessmode;
export const Systemusersaddress1_addresstypecode = {
  1: '既定値'
} as const;
export type Systemusersaddress1_addresstypecode = keyof typeof Systemusersaddress1_addresstypecode;
export const Systemusersaddress1_shippingmethodcode = {
  1: '既定値'
} as const;
export type Systemusersaddress1_shippingmethodcode = keyof typeof Systemusersaddress1_shippingmethodcode;
export const Systemusersaddress2_addresstypecode = {
  1: '既定値'
} as const;
export type Systemusersaddress2_addresstypecode = keyof typeof Systemusersaddress2_addresstypecode;
export const Systemusersaddress2_shippingmethodcode = {
  1: '既定値'
} as const;
export type Systemusersaddress2_shippingmethodcode = keyof typeof Systemusersaddress2_shippingmethodcode;
export const Systemusersazurestate = {
  0: '存在する',
  1: '論理的に削除済み',
  2: '見つからないか物理的に削除済み'
} as const;
export type Systemusersazurestate = keyof typeof Systemusersazurestate;
export const Systemuserscaltype = {
  0: 'プロフェッショナル',
  1: '管理',
  2: 'ベーシック',
  3: 'デバイス Professional',
  4: 'Device Basic',
  5: '必須',
  6: 'Device Essential',
  7: 'Enterprise',
  8: 'Device Enterprise',
  9: '営業',
  10: 'サービス',
  11: 'Field Service',
  12: 'Project Service'
} as const;
export type Systemuserscaltype = keyof typeof Systemuserscaltype;
export const Systemusersdeletedstate = {
  0: '削除されていない',
  1: '論理的に削除済み'
} as const;
export type Systemusersdeletedstate = keyof typeof Systemusersdeletedstate;
export const Systemusersemailrouteraccessapproval = {
  0: '空',
  1: '承認済み',
  2: '承認待ち',
  3: '却下済み'
} as const;
export type Systemusersemailrouteraccessapproval = keyof typeof Systemusersemailrouteraccessapproval;
export const Systemusersincomingemaildeliverymethod = {
  0: 'なし',
  1: 'Outlook 用 Microsoft Dynamics 365',
  2: 'サーバー側同期または E-mail Router',
  3: '転送用メールボックス'
} as const;
export type Systemusersincomingemaildeliverymethod = keyof typeof Systemusersincomingemaildeliverymethod;
export const Systemusersinvitestatuscode = {
  0: '未送信',
  1: '招待済み',
  2: '期限切れ間近',
  3: '期限切れ',
  4: '受諾',
  5: '拒否',
  6: '撤回'
} as const;
export type Systemusersinvitestatuscode = keyof typeof Systemusersinvitestatuscode;
export const Systemusersoutgoingemaildeliverymethod = {
  0: 'なし',
  1: 'Outlook 用 Microsoft Dynamics 365',
  2: 'サーバー側同期または E-mail Router'
} as const;
export type Systemusersoutgoingemaildeliverymethod = keyof typeof Systemusersoutgoingemaildeliverymethod;
export const Systemuserspreferredaddresscode = {
  1: '郵送先住所',
  2: '住所 (その他)'
} as const;
export type Systemuserspreferredaddresscode = keyof typeof Systemuserspreferredaddresscode;
export const Systemuserspreferredemailcode = {
  1: '既定値'
} as const;
export type Systemuserspreferredemailcode = keyof typeof Systemuserspreferredemailcode;
export const Systemuserspreferredphonecode = {
  1: '代表電話',
  2: 'その他の電話',
  3: '自宅電話',
  4: '携帯電話'
} as const;
export type Systemuserspreferredphonecode = keyof typeof Systemuserspreferredphonecode;
export const Systemuserssystemmanagedusertype = {
  0: 'Entra ユーザー',
  1: 'C2 ユーザー',
  2: '偽装可能なスタブ ユーザー',
  3: 'Agentic User',
  4: 'NonLicensed'
} as const;
export type Systemuserssystemmanagedusertype = keyof typeof Systemuserssystemmanagedusertype;

export interface SystemusersBase {
  accessmode: Systemusersaccessmode;
  address1_addressid?: string;
  address1_addresstypecode?: Systemusersaddress1_addresstypecode;
  address1_city?: string;
  address1_country?: string;
  address1_county?: string;
  address1_fax?: string;
  address1_latitude?: number;
  address1_line1?: string;
  address1_line2?: string;
  address1_line3?: string;
  address1_longitude?: number;
  address1_name?: string;
  address1_postalcode?: string;
  address1_postofficebox?: string;
  address1_shippingmethodcode?: Systemusersaddress1_shippingmethodcode;
  address1_stateorprovince?: string;
  address1_telephone1?: string;
  address1_telephone2?: string;
  address1_telephone3?: string;
  address1_upszone?: string;
  address1_utcoffset?: number;
  address2_addressid?: string;
  address2_addresstypecode?: Systemusersaddress2_addresstypecode;
  address2_city?: string;
  address2_country?: string;
  address2_county?: string;
  address2_fax?: string;
  address2_latitude?: number;
  address2_line1?: string;
  address2_line2?: string;
  address2_line3?: string;
  address2_longitude?: number;
  address2_name?: string;
  address2_postalcode?: string;
  address2_postofficebox?: string;
  address2_shippingmethodcode?: Systemusersaddress2_shippingmethodcode;
  address2_stateorprovince?: string;
  address2_telephone1?: string;
  address2_telephone2?: string;
  address2_telephone3?: string;
  address2_upszone?: string;
  address2_utcoffset?: number;
  applicationid?: string;
  azureactivedirectoryobjectid?: string;
  azurestate: Systemusersazurestate;
  "BusinessUnitId@odata.bind": string;
  "CalendarId@odata.bind"?: string;
  caltype: Systemuserscaltype;
  displayinserviceviews?: boolean;
  domainname: string;
  emailrouteraccessapproval: Systemusersemailrouteraccessapproval;
  employeeid?: string;
  entityimage?: string;
  firstname: string;
  governmentid?: string;
  homephone?: string;
  importsequencenumber?: number;
  incomingemaildeliverymethod: Systemusersincomingemaildeliverymethod;
  internalemailaddress: string;
  invitestatuscode: Systemusersinvitestatuscode;
  isallowedbyipfirewall?: boolean;
  isdisabled?: boolean;
  isintegrationuser: boolean;
  islicensed: boolean;
  issyncwithdirectory: boolean;
  jobtitle?: string;
  lastname: string;
  middlename?: string;
  mobilealertemail?: string;
  "MobileOfflineProfileId@odata.bind"?: string;
  mobilephone?: string;
  nickname?: string;
  outgoingemaildeliverymethod: Systemusersoutgoingemaildeliverymethod;
  overriddencreatedon?: string;
  "ParentSystemUserId@odata.bind"?: string;
  passporthi?: number;
  passportlo?: number;
  personalemailaddress?: string;
  photourl?: string;
  "PositionId@odata.bind"?: string;
  preferredaddresscode?: Systemuserspreferredaddresscode;
  preferredemailcode?: Systemuserspreferredemailcode;
  preferredphonecode?: Systemuserspreferredphonecode;
  processid?: string;
  "QueueId@odata.bind"?: string;
  salutation?: string;
  setupuser: boolean;
  sharepointemailaddress?: string;
  skills?: string;
  stageid?: string;
  systemmanagedusertype: Systemuserssystemmanagedusertype;
  systemuserid: string;
  "TerritoryId@odata.bind"?: string;
  timezoneruleversionnumber?: number;
  title?: string;
  "TransactionCurrencyId@odata.bind"?: string;
  traversedpath?: string;
  userlicensetype: number;
  utcconversiontimezonecode?: number;
  windowsliveid?: string;
  yammeremailaddress?: string;
  yammeruserid?: string;
  yomifirstname?: string;
  yomilastname?: string;
  yomimiddlename?: string;
}

export interface Systemusers extends SystemusersBase {
  accessmodename?: string;
  activedirectoryguid?: string;
  address1_addresstypecodename?: string;
  address1_composite?: string;
  address1_shippingmethodcodename?: string;
  address2_addresstypecodename?: string;
  address2_composite?: string;
  address2_shippingmethodcodename?: string;
  applicationiduri?: string;
  azuredeletedon?: string;
  azurestatename?: string;
  businessunitidname: string;
  caltypename?: string;
  createdbyname?: string;
  createdbyyominame?: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame?: string;
  defaultfilterspopulated: boolean;
  defaultmailboxname?: string;
  defaultodbfoldername: string;
  deletedstate: Systemusersdeletedstate;
  deletedstatename?: string;
  disabledreason?: string;
  displayinserviceviewsname?: string;
  emailrouteraccessapprovalname?: string;
  entityimage_timestamp?: number;
  entityimage_url?: string;
  entityimageid?: string;
  exchangerate?: number;
  fullname?: string;
  identityid: number;
  incomingemaildeliverymethodname?: string;
  invitestatuscodename?: string;
  isactivedirectoryuser: boolean;
  isallowedbyipfirewallname?: string;
  isdisabledname?: string;
  isemailaddressapprovedbyo365admin: boolean;
  isintegrationusername?: string;
  islicensedname?: string;
  latestupdatetime?: string;
  modifiedbyname?: string;
  modifiedbyyominame?: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame?: string;
  organizationid: string;
  organizationidname: string;
  outgoingemaildeliverymethodname?: string;
  parentsystemuseridname?: string;
  parentsystemuseridyominame?: string;
  positionidname?: string;
  preferredaddresscodename?: string;
  preferredemailcodename?: string;
  preferredphonecodename?: string;
  queueidname?: string;
  setupusername?: string;
  systemmanagedusertypename?: string;
  territoryidname: string;
  transactioncurrencyidname?: string;
  userpuid?: string;
  versionnumber?: number;
  yomifullname?: string;
  businessunitid?: object;
  _businessunitid_value?: string;
  calendarid?: object;
  _calendarid_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  defaultmailbox?: object;
  _defaultmailbox_value?: string;
  mobileofflineprofileid?: object;
  _mobileofflineprofileid_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  parentsystemuserid?: object;
  _parentsystemuserid_value?: string;
  positionid?: object;
  _positionid_value?: string;
  queueid?: object;
  _queueid_value?: string;
  territoryid?: object;
  _territoryid_value?: string;
  transactioncurrencyid?: object;
  _transactioncurrencyid_value?: string;
}