// apps/pc/src/generated/models/TeamsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Teamsmembershiptype = {
  0: 'メンバーとゲスト',
  1: 'メンバー',
  2: '所有者',
  3: 'ゲスト'
} as const;
export type Teamsmembershiptype = keyof typeof Teamsmembershiptype;
export const Teamsteamtype = {
  0: '所有者',
  1: 'アクセス',
  2: 'セキュリティ グループ',
  3: 'Office グループ'
} as const;
export type Teamsteamtype = keyof typeof Teamsteamtype;

export interface TeamsBase {
  "AdministratorId@odata.bind": string;
  azureactivedirectoryobjectid?: string;
  "BusinessUnitId@odata.bind": string;
  "DelegatedAuthorizationId@odata.bind"?: string;
  description?: string;
  emailaddress?: string;
  importsequencenumber?: number;
  membershiptype: Teamsmembershiptype;
  name: string;
  overriddencreatedon?: string;
  processid?: string;
  "QueueId@odata.bind"?: string;
  "RegardingObjectId@odata.bind"?: string;
  stageid?: string;
  teamid: string;
  "TeamTemplateId@odata.bind"?: string;
  teamtype: Teamsteamtype;
  "TransactionCurrencyId@odata.bind"?: string;
  traversedpath?: string;
  yominame?: string;
}

export interface Teams extends TeamsBase {
  administratoridname?: string;
  administratoridyominame?: string;
  businessunitidname: string;
  createdbyname?: string;
  createdbyyominame?: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame?: string;
  delegatedauthorizationidname?: string;
  exchangerate?: number;
  isdefault: boolean;
  isdefaultname?: string;
  issastokenset?: boolean;
  issastokensetname?: string;
  membershiptypename?: string;
  modifiedbyname?: string;
  modifiedbyyominame?: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame?: string;
  organizationid: string;
  organizationidname: string;
  queueidname?: string;
  regardingobjecttypecode?: string;
  sastoken?: string;
  sharelinkqualifier?: string;
  systemmanaged: boolean;
  systemmanagedname?: string;
  teamtypename?: string;
  transactioncurrencyidname?: string;
  versionnumber?: number;
  administratorid?: object;
  _administratorid_value?: string;
  businessunitid?: object;
  _businessunitid_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  delegatedauthorizationid?: object;
  _delegatedauthorizationid_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  queueid?: object;
  _queueid_value?: string;
  regardingobjectid?: object;
  _regardingobjectid_value?: string;
  teamtemplateid?: object;
  _teamtemplateid_value?: string;
  transactioncurrencyid?: object;
  _transactioncurrencyid_value?: string;
}